+++ ext/doc/Cassandra/DefaultCluster.php
@@ -22,17 +22,17 @@ namespace Cassandra;
 /**
  * Default cluster implementation.
  *
- * @see Cassandra::Cluster
+ * @see Cassandra\Cluster
  */
 final class DefaultCluster implements Cluster
 {
     /**
      * {@inheritDoc}
      */
-    public function Session connect(string $keyspace = null) {}
+    public function connect($keyspace = null) {}
 
     /**
      * {@inheritDoc}
      */
-    public function Future connectAsync(string $keyspace = null) {}
+    public function connectAsync($keyspace = null) {}
 }