+++ ext/doc/Cassandra/FutureClose.php
@@ -19,14 +19,14 @@
 namespace Cassandra;
 
 /**
- * A future returned from Cassandra::Session::closeAsync().
+ * A future returned from Cassandra\Session::closeAsync().
  *
- * @see Cassandra::Session::closeAsync()
+ * @see Cassandra\Session::closeAsync()
  */
 final class FutureClose implements Future
 {
     /**
      * {@inheritDoc}
      */
-    public function mixed get(float $timeout = null) {}
+    public function get($timeout = null) {}
 }