+++ ext/doc/Cassandra/FuturePreparedStatement.php
@@ -22,12 +22,12 @@ namespace Cassandra;
  * A future returned from `Cassandra\Session::prepareAsync()`
  * This future will resolve with a `Cassandra\PreparedStatement` or an exception.
  *
- * @see Cassandra::Session::prepareAsync()
+ * @see Cassandra\Session::prepareAsync()
  */
 final class FuturePreparedStatement implements Future
 {
     /**
      * {@inheritDoc}
      */
-    public function mixed get(float $timeout = null) {}
+    public function get($timeout = null) {}
 }