+++ ext/doc/Cassandra/FutureSession.php
@@ -19,14 +19,14 @@
 namespace Cassandra;
 
 /**
- * A future that resolves with `Cassandra::Session`.
+ * A future that resolves with `Cassandra\Session`.
  *
- * @see Cassandra::Cluster::connectAsync()
+ * @see Cassandra\Cluster::connectAsync()
  */
 final class FutureSession implements Future
 {
     /**
      * {@inheritDoc}
      */
-    public function mixed get(float $timeout = null) {}
+    public function get($timeout = null) {}
 }