+++ ext/doc/Cassandra/FutureValue.php
@@ -26,5 +26,5 @@ final class FutureValue implements Future
     /**
      * {@inheritDoc}
      */
-    public function mixed get(float $timeout = null) {}
+    public function get($timeout = null) {}
 }