+++ ext/doc/Cassandra/Set.php
@@ -18,50 +18,47 @@
 
 namespace Cassandra;
 
-use Countable;
-use IteratorAggregate;
-
 /**
  * A PHP representation of the CQL `set` datatype
  */
-final class Set implements Countable, IteratorAggregate
+final class Set implements \Countable, \IteratorAggregate
 {
     /**
      * Creates a new collection of a given type
      * @param string $type one of `Cassandra::TYPE_*`
      */
-    public function __construct(string $type) {}
+    public function __construct($type) {}
 
     /**
      * The type of this set
      * @return string one of `Cassandra::TYPE_*`
      */
-    public function string type() {}
+    public function type() {}
 
     /**
      * Array of values in this set
      * @return array values
      */
-    public function array values() {}
+    public function values() {}
 
     /**
      * Adds a value to this set
      * @param  mixed   $value Value
      * @return bool whether the value has been added
      */
-    public function bool add(mixed $value) {}
+    public function add(mixed $value) {}
 
     /**
      * Removes a value to this set
      * @param  mixed   $value Value
      * @return bool whether the value has been removed
      */
-    public function bool remove(mixed $value) {}
+    public function remove(mixed $value) {}
 
     /**
      * Returns whether a value is in this set
      * @param  mixed   $value Value
      * @return bool whether the value is in the set
      */
-    public function bool has(mixed $value) {}
+    public function has(mixed $value) {}
 }