+++ ext/doc/Cassandra/Timestamp.php
@@ -18,8 +18,6 @@
 
 namespace Cassandra;
 
-use DateTime;
-
 /**
  * A PHP representation of the CQL `timestamp` datatype
  */
@@ -31,14 +29,14 @@ final class Timestamp
      * @param int $time Unix timestamp
      * @param int $usec Microseconds
      */
-    public function __construct(int $time = null, int $usec = null) {}
+    public function __construct($time = null, $usec = null) {}
 
     /**
      * Unix timestamp
      * @see time
      * @return int seconds
      */
-    public function int time() {}
+    public function time() {}
 
     /**
      * Microtime from this timestamp
@@ -46,13 +44,13 @@ final class Timestamp
      * @param  bool      $get_as_float Whther to get this value as float
      * @return float|string               Float or string representation
      */
-    public function float|string microtime(bool $get_as_float = false) {}
+    public function microtime($get_as_float = false) {}
 
     /**
      * Converts current timestamp to PHP DateTime
      * @return DateTime
      */
-    public function DateTime toDateTime() {}
+    public function toDateTime() {}
 
     /**
      * Returns a string representation of this timestamp