+++ ext/doc/Cassandra/Timeuuid.php
@@ -18,8 +18,6 @@
 
 namespace Cassandra;
 
-use DateTime;
-
 /**
  * A PHP representation of the CQL `timeuuid` datatype
  */
@@ -29,36 +27,36 @@ final class Timeuuid
      * Creates a timeuuid from a given timestamp or current time.
      * @param int $timestamp Unix timestamp
      */
-    public function __construct(int $timestamp = null) {}
+    public function __construct($timestamp = null) {}
 
     /**
      * Returns this timeuuid as string
      * @return string timeuuid
      */
-    public function string __toString() {}
+    public function __toString() {}
 
     /**
      * Returns this timeuuid as string
      * @return string timeuuid
      */
-    public function string uuid() {}
+    public function uuid() {}
 
     /**
      * Returns the version of this timeuuid
      * @return int version of this timeuuid
      */
-    public function int version() {}
+    public function version() {}
 
     /**
      * Unix timestamp
      * @see time()
      * @return int seconds
      */
-    public function int time() {}
+    public function time() {}
 
     /**
      * Converts current timeuuid to PHP DateTime
      * @return DateTime
      */
-    public function DateTime toDateTime() {}
+    public function toDateTime() {}
 }