+++ ext/src/Cassandra/Varint.c
@@ -61,11 +61,11 @@ PHP_METHOD(Varint, __construct)
 
   self = (cassandra_varint*) zend_object_store_get_object(getThis() TSRMLS_CC);
 
-  if(Z_TYPE_P(num) == IS_LONG) {
+  if (Z_TYPE_P(num) == IS_LONG) {
     mpz_set_si(self->value, Z_LVAL_P(num));
-  } else if(Z_TYPE_P(num) == IS_DOUBLE) {
+  } else if (Z_TYPE_P(num) == IS_DOUBLE) {
     mpz_set_d(self->value, Z_DVAL_P(num));
-  } else if(Z_TYPE_P(num) == IS_STRING) {
+  } else if (Z_TYPE_P(num) == IS_STRING) {
     php_cassandra_parse_varint(Z_STRVAL_P(num), Z_STRLEN_P(num), &self->value TSRMLS_CC);
   } else if (Z_TYPE_P(num) == IS_OBJECT &&
              instanceof_function(Z_OBJCE_P(num), cassandra_varint_ce TSRMLS_CC)) {