+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 1.0.0-beta
+
+Features:
+
+* Persistent Sessions (can be disabled via `Cassandra\Cluster\Builder::withPersistentSessions()`).
+* Windows support.
+* Configurable log location (via `cassandra.log` ini setting).
+* Configurable log level (via `cassandra.log_level` ini setting).
+* Upgraded to CPP driver v2.0.
+* Added math functions to numeric types and `Cassandra\Numeric` interface.
+* All classes moved into the extension.
+
+Bug Fixes:
+
+* [PHP-18] Decimal parsing failure
+
 # 1.0.0-alpha
 
 Features: